Organyze_backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.10           # via django
django-cors-headers==3.5.0  # via -r requirements.in
django==3.1.2             # via django-cors-headers, djangorestframework
djangorestframework-camel-case==1.2.0  # via -r requirements.in
djangorestframework==3.12.1  # via -r requirements.in
markdown==3.3.1           # via -r requirements.in
psycopg2-binary==2.8.6    # via -r requirements.in
pytz==2020.1              # via django
sqlparse==0.4.1           # via django
